# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/rol/{id}": {
      "get": {
        "description": "Retrieves a role by its unique identifier",
        "operationId": "RolController_getRolById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Role ID",
            "schema": {
              "example": "rol-123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rol"
                }
              }
            }
          },
          "404": {
            "description": "Role not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get role by ID",
        "tags": [
          "rol"
        ]
      }
    },
    "/rol/name/{name}": {
      "get": {
        "description": "Retrieves a role by its name",
        "operationId": "RolController_getRolByName",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Role name",
            "schema": {
              "example": "admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rol"
                }
              }
            }
          }
        },
        "summary": "Get role by name",
        "tags": [
          "rol"
        ]
      }
    },
...
    "/module/{id}": {
      "get": {
        "description": "Retrieves a module by its unique identifier",
        "operationId": "ModuleController_getModuleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Module ID",
            "schema": {
              "example": "module-123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Module found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Module"
                }
              }
            }
          },
          "404": {
            "description": "Module not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get module by ID",
        "tags": [
          "module"
        ]
      },
      "patch": {
        "description": "Updates module configuration and status",
        "operationId": "ModuleController_updateModuleConfigurations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Module ID",
            "schema": {
              "example": "module-123"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Module"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Module updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Module"
                }
              }
            }
          },
          "404": {
            "description": "Module not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update module configurations",
        "tags": [
          "module"
        ]
      }
    },
...
    "/feature/{id}": {
      "get": {
        "description": "Retrieves a feature by its unique identifier",
        "operationId": "FeatureController_getFeatureToggleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Feature ID",
            "schema": {
              "example": "feature-123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              }
            }
          },
          "404": {
            "description": "Feature not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get feature by ID",
        "tags": [
          "feature"
        ]
      }
    },
    "/feature": {
      "post": {
        "description": "Creates a new system feature definition",
        "operationId": "FeatureController_createFeature",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "new-feature"
                  },
                  "description": {
                    "type": "string",
                    "example": "Feature description"
                  },
                  "type": {
                    "type": "string",
                    "example": "MODULE"
                  },
                  "module": {
                    "type": "string",
                    "example": "module-name"
                  }
                },
                "required": [
                  "name",
                  "type"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feature created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new feature",
        "tags": [
          "feature"
        ]
      }
    },
    "/feature/name/{name}": {
      "get": {
        "description": "Retrieves a feature by its name",
        "operationId": "FeatureController_getFeatureToggleByName",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Feature name",
            "schema": {
              "example": "new-feature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              }
            }
          }
        },
        "summary": "Get feature by name",
        "tags": [
          "feature"
        ]
      }
    },
    "/feature/type/{type}": {
      "get": {
        "description": "Retrieves all features of a specific type",
        "operationId": "FeatureController_getFeatureToggleByType",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "description": "Feature type",
            "schema": {
              "example": "MODULE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of features",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feature"
                  }
                }
              }
            }
          }
        },
        "summary": "Get features by type",
        "tags": [
          "feature"
        ]
      }
    },
    "/feature/module/{module}": {
      "get": {
        "description": "Retrieves all features for a specific module",
        "operationId": "FeatureController_getFeatureByModule",
        "parameters": [
          {
            "name": "module",
            "required": true,
            "in": "path",
            "description": "Module name",
            "schema": {
              "example": "module-name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of features",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feature"
                  }
                }
              }
            }
          }
        },
        "summary": "Get features by module",
        "tags": [
          "feature"
        ]
      }
    },
...
  },
  "info": {
    "title": "ESV Backend API",
    "description": "API documentation for ESV Backend - Appointment and Queue Management System",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://app.easuavez.com",
      "description": "Default server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "Rol": {
        "type": "object",
        "properties": {}
      },
...
      "Module": {
        "type": "object",
        "properties": {}
      },
...
      "Feature": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}